Do not autoload session data.
Session data should only be accessed from the database if the user that that data belongs to is visiting the site. Since "update_option" defaults to autoloading options, the option is added via "add_option" when the session data is initially set to the database. This should allow only the current user's session data to load, instead of loading all users' session data.

Fixes #6