Introduce Recursive_ArrayAccess
The `Recursive_ArrayAccess` class allows for multidimensional arrays
within the `WP_Session` object, fixing an issue reported on the WP
support forums.  Additionally, this change introduces the concept of a
session expiration variant (a second expiration time).  If the current
time is less than the variant, then no expiration times are updated in
the database.  Once the time passes the variant, the expiration time for
the session is reset to default values.  This prevents needlessly
writing to the database one every page load.  Also, the session itself
will only be written to the database if it has been changed and flagged
as dirty.  These two changes fix issue #4.