Initial 2.0 Refactoring
Instead of managing all of the session data and functionality directly,
use Sessionz as an abstraction so we can leverage _native_ functionality
within PHP itself.

The Sessionz stack will store sessions in:
- Memory (unencrypted)
- falling back to the Options table (where things are encrypted)

Storing the integer timestamp for when a session was created in the
'expiration' option makes for easier cleanup and simplifies the database
queries that need to run.

Existing hooks and functions from the 1.X branch are retained, but will
trigger a WordPress "doing it wrong" error when invoked as these should
not be used moving forward.