Change prefix for better integration with proxies
The `wordpress_` prefix for is commonly being used for wordpress internal
cookies. Some wordpress hosting sites, for example wpengine, are using
those cookies to decide if a page will be served from cache or from the
application server directly. I propose to rename the the wp_session
cookie to conform this prefix as a saner default.