Avoid storing empty session in options table
We have a website that is storing 250k-500k options as a result of this library being included -- even with frequent garbage collection -- as a new session is created for every visitor regardless of whether it is needed to store anything. As WordPress has a "default" argument for get_option, there's no reason to clutter up the options table with empty strings or empty arrays.